fix: Fix acceptance tests
The path was in the old URL schema, which we do not really seem to
support for new routes.

eg: /organizations/FOO/settings/ redirects to /settings/FOO/
but it's not the same for our new settings page

Also the entire setUp block was completely useless. No test was using
that.